+++ spec/spec_helper.rb
@@ -44,4 +44,9 @@ RSpec.configure do |config|
   config.example_status_persistence_file_path = '.rspec_status'
   config.disable_monkey_patching!
   config.expect_with(:rspec) { |c| c.syntax = :expect }
+
+  config.before do
+    Legion::Extensions::PilotInfraMonitor::StateTracker.reset!
+    Legion::Extensions::PilotInfraMonitor::AlertDedup.reset!
+  end
 end